+++ gradle.properties
@@ -1,4 +1,4 @@
-VERSION_NAME=6.0.0
+VERSION_NAME=6.1.0
 VERSION_CODE=600000
 GROUP=com.karumi
 POM_DESCRIPTION=Gradle plugin developed to facilitate screenshot testing for Android.
